backend/app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-anthropic>=0.3.15",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
]
